Adding new Vunlerability Assessment Baseline API
Hi

We, the SQL Security team in ILDC, have issued a private preview new
offering for SQL DB named Vulnerability Assessment.

Using the Vulnerability Assessment offering, customers can scan their
databases with a predefined set of rules (security checks) and get a
full picture of their SQL vulnerabilities and have the ability to
remediate each one.
(More info can be found at Blog - New Vulnerability Assessment Service
for Azure SQL DB and 1-pager - VulnerabilityAssessment)

We are currently adding a new feature of Baseline, where a customer can
mark a certain failure as passing in his environment.

An example of using this feature can be for a rule of “No user-defined
database roles are members of fixed roles” where in a certain customer
environment exists a power user that is a member of fixed role.
When adding this result to the rule baseline, this check will not fail.