Add swagger for continuous export configuration and features and pricing apis.
1.	Rename appInsights to applicaitonInsights because appinsights is
a brand of another company, and we are using applicaitoninsights in all
documents.
2.	Add swagger for continuous export configuration apis, it include
get all export configurations for a component, create a new export
configuration, get a specific export configuration, delete a specific
export configuration, update a specific export configuration.
3.	Add swagger for features and pricing, it include get current
billing features for a component, update billing features for a
component, and get back current daily volume cap status for a component.